backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Werkzeug==3.0.4
python-dotenv==1.2.2
